Docker

+++ handlers/characters.js
@@ -1045,11 +1045,12 @@ function embedForCharacter(msg, charArray, title, isShow, vaultUser) {
         }
         // console.log('vaultuser', vaultUser);
         let defCharString = vaultUser?.defaultCharacter == char.id ? ` ${utils.EMOJIS.ASTERISK}` : '';
+        let charNameString = isShow ? `[${char.name}](${char.readonlyUrl})${defCharString}` : stringForCharacter(char);
         // console.log('defCharString "%s" and "%s"', defCharString, char.id);
         charEmbed.addFields(
             {
                 name: `\:dagger: Name | ID | Status | Campaign \:shield:`,
-                value: `[${char.name}](${char.readonlyUrl})${defCharString} | ${char.id} |
+                value: `${charNameString} | ${char.id} |
                     ${stringForApprovalsAndUpdates(char)} | ${stringForCampaign(char)}`
             }
         );
@@ -1330,11 +1331,11 @@ async function handleShow(msg, msgParms, guildConfig) {
  */
 async function handleCampaign(msg, msgParms, guildConfig) {
     try {
-        if (msgParms.length < 2) {
-            throw new Error('Please pass the character id and the campaign id.');
+        if (msgParms.length < 1) {
+            throw new Error('Please pass the (in the least) the character id.');
         }
-        const charID = msgParms[0].value;
-        const campaignID = msgParms[1].value;
+        const charID = msgParms[0]?.value;
+        const campaignID = msgParms[1]?.value;
         // console.log(`charid: ${charID} campaignID: ${campaignID}`);
 
         const charToEdit = await CharModel.findOne({ guildUser: msg.member.id, id: charID, isUpdate: false, guildID: msg.guild.id });
@@ -1343,8 +1344,10 @@ async function handleCampaign(msg, msgParms, guildConfig) {
         }
         charToEdit.campaignOverride = campaignID;
         await charToEdit.save();
-        await utils.sendDirectOrFallbackToChannel({ name: 'Campaign', value: `Character ${charID} campaign changed to ${campaignID}` }, msg);
-        await msg.delete();
+        await utils.sendDirectOrFallbackToChannel({ name: 'Campaign', value: `Character ${charID} campaign changed to \`${campaignID ? campaignID : 'No Campaign Override'}\`` }, msg);
+        if (msg.deletable) {
+            await msg.delete();
+        }
     } catch (error) {
         await utils.sendDirectOrFallbackToChannelError(error, msg);
     }
